Fog of war now obscures units
Changes to how mapDrawer renders, has two lists of things to draw one
list is always visible the other relies on player los. Map is divided up
into tiles, if you have a unit within los distance of the centre of a
tile you can see that tile and everything in it.